BaseItem, BaseBlock, BaseMachine are all now abstract classes.
Removed test items and test blocks, commented for reference.

Proxy calls 2 functions now, the initial and the recipes, this allows
for recipes to be located inside the respective block or item class
for visibility.

Added a test recipe of an in-mod item (Bubba Item) to create in-mod
block (Coal Generator).